Alterando de INTEGER para NUMERIC

+++ biblioteca.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "funcionarios_biblioteca" (
-  "matricula" INTEGER NOT NULL,
-  "codigo_unidade" INTEGER NULL CHECK(
+  "matricula" NUMERIC(5) NOT NULL,
+  "codigo_unidade" NUMERIC(3) NULL CHECK(
     ("tipo_funcionario" = 'A' AND "codigo_unidade" IS NOT NULL) OR
     ("tipo_funcionario" = 'B' AND "codigo_unidade" IS NULL)
   ),
@@ -11,8 +11,8 @@ CREATE TABLE "funcionarios_biblioteca" (
 );
 
 CREATE TABLE "unidades_atendimento" (
-  "codigo" INTEGER NOT NULL,
-  "matricula_bibliotecaria" INTEGER NOT NULL UNIQUE  --Verificar se a matricula é realmente de uma bibliotecaria,
+  "codigo" NUMERIC(3) NOT NULL,
+  "matricula_bibliotecaria" NUMERIC(5) NOT NULL UNIQUE  --Verificar se a matricula é realmente de uma bibliotecaria,
   "nome" VARCHAR(50) NOT NULL UNIQUE,
   "endereco" VARCHAR(200) NOT NULL,
   PRIMARY KEY("codigo"),
@@ -20,8 +20,8 @@ CREATE TABLE "unidades_atendimento" (
 );
 
 CREATE TABLE "telefones_unidades" (
-  "codigo_unidade" INTEGER NOT NULL,
-  "telefone" INTEGER NOT NULL,
+  "codigo_unidade" NUMERIC(3) NOT NULL,
+  "telefone" NUMERIC(10) NOT NULL,
   PRIMARY KEY("codigo_unidade", "telefone"),
   FOREIGN KEY("codigo_unidade") REFERENCES "unidades_atendimento"("codigo")
 );
@@ -33,7 +33,7 @@ CREATE TABLE "unidades_academicas" (
 );
 
 CREATE TABLE "cursos" (
-  "codigo" INTEGER NOT NULL,
+  "codigo" NUMERIC(3) NOT NULL,
   "nome" VARCHAR(50) NOT NULL UNIQUE,
   "codigo_unidade" CHAR(4) NOT NULL,
   PRIMARY KEY("codigo"),
@@ -47,17 +47,17 @@ CREATE TABLE "disciplinas" (
 );
 
 CREATE TABLE "titulos" (
-  "isbn" INTEGER NOT NULL,
+  "isbn" NUMERIC(5) NOT NULL,
   "nome_titulo" VARCHAR(100) NOT NULL,
   "area_principal" VARCHAR(100) NOT NULL,
   "assunto" VARCHAR(100) NOT NULL,
-  "ano_publicacao" INTEGER NOT NULL,
+  "ano_publicacao" NUMERIC(4) NOT NULL,
   "editora" VARCHAR(50) NOT NULL,
   "idioma" CHAR(1) NOT NULL,
-  "prazo_emprestimo_professor" INTEGER NOT NULL,
-  "prazo_emprestimo_aluno" INTEGER NOT NULL,
-  "numero_max_renovacao" INTEGER NOT NULL,
-  "edicao" INTEGER NULL,
+  "prazo_emprestimo_professor" NUMERIC(2) NOT NULL,
+  "prazo_emprestimo_aluno" NUMERIC(2) NOT NULL,
+  "numero_max_renovacao" NUMERIC(2) NOT NULL,
+  "edicao" NUMERIC(3) NULL,
   "periodicidade" CHAR(15) NULL CHECK("periodicidade" IN ('SEMANAL', 'QUINZENAL', 'MENSAL', 'TRIMESTRAL', 'QUADRIMESTRAL', 'SEMESTRAL', 'ANUAL')),
   "tipo_periodico" CHAR(1) NULL CHECK("tipo_periodico" IN ('J', 'R', 'B')),
   "tipo_titulo" CHAR(1) NOT NULL CHECK("tipo_titulo" IN ('L', 'P')),
@@ -69,21 +69,21 @@ CREATE TABLE "titulos" (
 );
 
 CREATE TABLE "autores" (
-  "isbn_titulo" INTEGER NOT NULL,
+  "isbn_titulo" NUMERIC(5) NOT NULL,
   "autor" VARCHAR(100) NOT NULL,
   PRIMARY KEY("isbn_titulo", "autor"),
   FOREIGN KEY("isbn_titulo") REFERENCES "titulos"("isbn")
 );
 
 CREATE TABLE "areas_secundarias" (
-  "isbn_titulo" INTEGER NOT NULL,
+  "isbn_titulo" NUMERIC(5) NOT NULL,
   "area_secundaria" VARCHAR(100) NOT NULL,
   PRIMARY KEY("isbn_titulo","area_secundaria"),
   FOREIGN KEY("isbn_titulo") REFERENCES "titulos"("isbn")
 );
 
 CREATE TABLE "usuarios_biblioteca" (
-  "codigo" INTEGER NOT NULL,
+  "codigo" NUMERIC(5) NOT NULL,
   "codigo_unidade" CHAR(4) NULL CHECK("tipo_usuario" = 'P' AND "codigo_unidade" IS NOT NULL),
   "nome" VARCHAR(50) NOT NULL,
   "identidade" CHAR(12) NULL,
@@ -92,7 +92,7 @@ CREATE TABLE "usuarios_biblioteca" (
   "sexo" CHAR(1) NOT NULL CHECK("sexo" IN ('M', 'F')),
   "data_nascimento" DATE NOT NULL,
   "estado_civil" CHAR(1) NOT NULL CHECK("estado_civil" IN ('C', 'S', 'D', 'V')),
-  "matricula_professor" INTEGER NULL UNIQUE CHECK(
+  "matricula_professor" NUMERIC(5) NULL UNIQUE CHECK(
     ("tipo_usuario" = 'P' AND "matricula_professor" IS NOT NULL) OR
     ("tipo_usuario" = 'A' AND "matricula_professor" IS NULL)
   ),
@@ -102,16 +102,16 @@ CREATE TABLE "usuarios_biblioteca" (
 );
 
 CREATE TABLE "telefones_usuarios" (
-  "codigo_usuario" INTEGER NOT NULL,
-  "telefone" INTEGER NOT NULL,
+  "codigo_usuario" NUMERIC(5) NOT NULL,
+  "telefone" NUMERIC(10) NOT NULL,
   PRIMARY KEY("codigo_usuario", "telefone"),
   FOREIGN KEY("codigo_usuario") REFERENCES "usuarios_biblioteca"("codigo")
 );
 
 CREATE TABLE "transacoes" (
-  "numero_transacao" INTEGER NOT NULL,
-  "codigo_usuario" INTEGER NOT NULL,
-  "matricula_atendente" INTEGER NOT NULL --Verificar se a matricula é realmente de uma atendente,
+  "numero_transacao" NUMERIC(9) NOT NULL,
+  "codigo_usuario" NUMERIC(5) NOT NULL,
+  "matricula_atendente" NUMERIC(5) NOT NULL --Verificar se a matricula é realmente de uma atendente,
   "data_transacao" DATE NOT NULL DEFAULT CURRENT_DATE,
   "horario_transacao" TIME NOT NULL DEFAULT CURRENT_TIME,
   "tipo_transacao" CHAR(10) NOT NULL CHECK("tipo_transacao" IN ('EMPRESTIMO', 'DEVOLUÇÃO', 'RENOVAÇÃO', 'RESERVA')),
@@ -121,11 +121,11 @@ CREATE TABLE "transacoes" (
 );
 
 CREATE TABLE "copias_titulos" (
-  "numero_copia" INTEGER NOT NULL,
-  "isbn_titulo" INTEGER NOT NULL,
-  "codigo_unidade" INTEGER NOT NULL, 
-  "secao_copia" INTEGER NOT NULL,
-  "estante_copia" INTEGER NOT NULL,
+  "numero_copia" NUMERIC(5) NOT NULL,
+  "isbn_titulo" NUMERIC(5) NOT NULL,
+  "codigo_unidade" NUMERIC(3) NOT NULL, 
+  "secao_copia" NUMERIC(4) NOT NULL,
+  "estante_copia" NUMERIC(3) NOT NULL,
   PRIMARY KEY("numero_copia", "isbn_titulo"),
   FOREIGN KEY("isbn_titulo") REFERENCES "titulos"("isbn"),
   FOREIGN KEY("codigo_unidade") REFERENCES "unidades_atendimento"("codigo")
@@ -133,10 +133,10 @@ CREATE TABLE "copias_titulos" (
 );
 
 CREATE TABLE "itens_emprestimo" (
-  "numero_item" INTEGER NOT NULL,
-  "numero_transacao" INTEGER NOT NULL,
-  "numero_copia" INTEGER NOT NULL,
-  "isbn_titulo" INTEGER NOT NULL,
+  "numero_item" NUMERIC(1) NOT NULL,
+  "numero_transacao" NUMERIC(9) NOT NULL,
+  "numero_copia" NUMERIC(5) NOT NULL,
+  "isbn_titulo" NUMERIC(5) NOT NULL,
   "data_limite_devolucao" DATE NOT NULL,
   PRIMARY KEY("numero_item", "numero_transacao"),
   FOREIGN KEY("numero_transacao") REFERENCES "transacoes"("numero_transacao"),
@@ -145,9 +145,9 @@ CREATE TABLE "itens_emprestimo" (
 );
 
 CREATE TABLE "cursos_aluno" (
-  "codigo_curso" INTEGER NOT NULL,
-  "codigo_aluno" INTEGER NOT NULL --Verificar se o código é realmente de um aluno,
-  "matricula" INTEGER NOT NULL UNIQUE,
+  "codigo_curso" NUMERIC(3) NOT NULL,
+  "codigo_aluno" NUMERIC(5) NOT NULL --Verificar se o código é realmente de um aluno,
+  "matricula" NUMERIC(5) NOT NULL UNIQUE,
   PRIMARY KEY("codigo_curso", "codigo_aluno"),
   FOREIGN KEY("codigo_curso") REFERENCES "cursos"("codigo"),
   FOREIGN KEY("codigo_aluno") REFERENCES "usuarios_biblioteca"("codigo")
@@ -155,7 +155,7 @@ CREATE TABLE "cursos_aluno" (
 
 CREATE TABLE "disciplinas_professor" (
   "codigo_disciplina" CHAR(7) NOT NULL,
-  "matricula_professor" INTEGER NOT NULL,
+  "matricula_professor" NUMERIC(5) NOT NULL,
   PRIMARY KEY("codigo_disciplina", "matricula_professor"),
   FOREIGN KEY("codigo_disciplina") REFERENCES "disciplinas"("codigo")
   FOREIGN KEY("matricula_professor") REFERENCES "usuarios_biblioteca"("matricula_professor")
